Use jqplot for the line graphs, bar and pie charts
This change introduces the necessary code for using jqplot as the
charting and graphing library.

There is currently a problem with jqplot with dynamically created
elements, such that we had to create an element in the HTML DOM,
chartdiv-wrapper, and we then move the chart to the correct location.
This has the downside that smaller charts are not properly downscaled
(such as in the calendar-views in "Visitors this month").

Also, the default colours are picked from the screen, which might not be
as accurate as using the values from the theme.

If you want to compare the jqplot-version with the original swf-version,
the console gives you some pointers on how to change back to the flash
graph.